requirements.txt
python-telegram-bot==20.7
asyncpg==0.29.0
reportlab==4.0.7
python-dotenv==1.0.0